Refactor for deploy to PCF
