+++ manifest.yml
@@ -1,0 +1,8 @@
+---
+applications:
+  - name: gmdb-movie-svc
+    buildpacks: 
+      - java_buildpack
+    path: build/libs/gmdb-movie-svc-0.0.1-SNAPSHOT.jar
+    services:
+      - gmdb-devdb
