+++ settings.gradle
@@ -1,1 +1,1 @@
-rootProject.name = 'gmdb-movies-svc'
+rootProject.name = 'gmdb-movie-svc'
