+++ src/main/java/com/galvanize/gmdb/movie/GmdbMovieSvcApplication.java
@@ -1,0 +1,13 @@
+package com.galvanize.gmdb.movie;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class GmdbMovieSvcApplication {
+
+    public static void main(String[] args) {
+        SpringApplication.run(GmdbMovieSvcApplication.class, args);
+    }
+
+}
