+++ src/main/resources/application.properties
@@ -1,19 +1,22 @@
-# Runtime Configs
-EUREKA_HOST=gmdb-discovery:8761
-DATABASE_HOST=gmdb-devdb:3306
+debug=false
+
+# Deploy Props
+DB_HOST=localhost
+DB_PORT=3306
 DB_USER=gmdb_app
-DB_PASSWORD=1234
+DB_PWD=1234
 
 # Application props
-spring.application.name=gmdb-movies
-eureka.client.service-url.defaultZone=http://${EUREKA_HOST}/eureka
-eureka.client.register-with-eureka=true
 server.port=8200
-debug=false
-spring.datasource.url=jdbc:mysql://${DATABASE_HOST}/gmdb?serverTimezone=UTC&useSSL=false
+spring.application.name=gmdb-movie-svc
+
+# Database Props
+spring.datasource.url=jdbc:mysql://${DB_HOST}:${DB_PORT}/gmdb?useSSL=false
 spring.datasource.username=${DB_USER}
-spring.datasource.password=${DB_PASSWORD}
+spring.datasource.password=${DB_PWD}
 spring.datasource.driver-class-name=com.mysql.jdbc.Driver
+
+# JPA Props
 spring.jpa.hibernate.ddl-auto=update
-spring.jpa.generate-ddl=true
+spring.jpa.generate-ddl=false
 spring.jpa.database-platform=org.hibernate.dialect.MySQL57Dialect
